quick fix

+++ database/migrations/2022_09_08_083446_create_events_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->time('time');
             $table->date('date');
             $table->string('location');
-            $table->integer('userCreated');
+            $table->bigInteger('userCreated');
             $table->timestamps();
         });
     }
